ObjectPooler
Update: When adding an item to the pool, call the factory to retrieve the key saved during the prefab loading process with Addressables.

Reason: This is because the pool's key is a type of GameObject, while the key in the factory is the GameObject used as the key in the pool, meaning it was added first.